staging: comedi: add CMDF_BOGUS as synonym for TRIG_BOGUS
`TRIG_BOGUS` is one of the values that can be OR-ed into the `flags`
member of `struct comedi_cmd`.  It causes the comedi core to go through
some of the motions of setting up an asynchronous command without
actually setting it up.  If all goes well, it causes the command set-up
to fail with an `EAGAIN` error.

The name `TRIG_BOGUS` is a bit of a left-over from earlier times.  A
better name would be `CMDF_BOGUS` to match various other command flags.

Define `CMDF_BOGUS` with the same numeric value as the old `TRIG_BOGUS`
and redefine `TRIG_BOGUS` as a synonym of it.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>
Reviewed-by: H Hartley Sweeten <hsweeten@visionengravers.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>